Changing coordinates to integers

+++ src/main/java/pwr/tiirt/energy/saver/Antenna.java
@@ -5,12 +5,12 @@ import java.util.List;
 
 public class Antenna {
 
-	public final double x;
-	public final double y;
+	public final int x;
+	public final int y;
 	public final List<Antenna> neighbours;
 	public final boolean active;
 
-	public Antenna(final double x, final double y, final List<Antenna> neighbours, final boolean active) {
+	public Antenna(final int x, final int y, final List<Antenna> neighbours, final boolean active) {
 		this.x = x;
 		this.y = y;
 		this.neighbours = Collections.unmodifiableList(neighbours);
